+++ src/inmanta/data/__init__.py
@@ -807,7 +807,8 @@ class Environment(BaseDocument):
             name=AUTOSTART_AGENT_DEPLOY_INTERVAL,
             typ="int",
             default=600,
-            doc="The deployment interval of the autostarted agents.",
+            doc="The deployment interval of the autostarted agents."
+            " See also: :inmanta.config:option:`config.agent-deploy-interval`",
             validator=convert_int,
             agent_restart=True,
         ),
@@ -815,7 +816,8 @@ class Environment(BaseDocument):
             name=AUTOSTART_AGENT_DEPLOY_SPLAY_TIME,
             typ="int",
             default=10,
-            doc="The splay time on the deployment interval of the autostarted agents.",
+            doc="The splay time on the deployment interval of the autostarted agents."
+            " See also: inmanta.config:option:`config.agent-deploy-splay-time`",
             validator=convert_int,
             agent_restart=True,
         ),
@@ -823,7 +825,7 @@ class Environment(BaseDocument):
             name=AUTOSTART_AGENT_REPAIR_INTERVAL,
             typ="int",
             default=86400,
-            doc="The repair interval of the autostarted agents.",
+            doc="The repair interval of the autostarted agents. See also: inmanta.config:option:`config.agent-repair-interval`",
             validator=convert_int,
             agent_restart=True,
         ),
@@ -831,7 +833,8 @@ class Environment(BaseDocument):
             name=AUTOSTART_AGENT_REPAIR_SPLAY_TIME,
             typ="int",
             default=600,
-            doc="The splay time on the repair interval of the autostarted agents.",
+            doc="The splay time on the repair interval of the autostarted agents."
+            " See also: inmanta.config:option:`config.agent-repair-splay-time`",
             validator=convert_int,
             agent_restart=True,
         ),
@@ -848,7 +851,7 @@ class Environment(BaseDocument):
             typ="dict",
             validator=convert_agent_map,
             doc="A dict with key the name of agents that should be automatically started. The value "
-            "is either an empty string or an agent map string.",
+            "is either an empty string or an agent map string. See also :inmanta.config:option:`config.agent-map`",
             agent_restart=True,
         ),
         AUTOSTART_AGENT_INTERVAL: Setting(